requirements.txt
kafka-python
pandas
requests
s3fs
awscli